added one component
